fix: userId int -> long으로 타입 변경

+++ tellme/src/main/java/com/tellme/tellme/domain/survey/persistence/SurveyShortUrlRepository.java
@@ -8,5 +8,5 @@ import org.springframework.stereotype.Repository;
 public interface SurveyShortUrlRepository extends JpaRepository<SurveyShortUrl, Integer> {
     SurveyShortUrl findByUrl(String url);
 
-    SurveyShortUrl findBySurveyIdAndUserId(int surveyId, long userId);
+    SurveyShortUrl findBySurveyIdAndUserId(int surveyId, int userId);
 }
